wip: adds a root component for courses.
thinning down the controller makes this a bit less unwieldy to work
with, and will set us up for further refactorings.
