+++ app/styles/components/courses/root.scss
@@ -1,7 +1,7 @@
-@use "../ilios-common/mixins" as cm;
-@use "../mixins" as m;
+@use "../../ilios-common/mixins" as cm;
+@use "../../mixins" as m;
 
-.courses-list {
+.courses-root {
   @include cm.main-section;
 
   .filters {
@@ -36,14 +36,6 @@
       .actions {
         @include m.main-list-box-header-actions;
       }
-    }
-  }
-
-  .list {
-    @include m.main-list-box-table;
-
-    i {
-      margin-left: 0.5rem;
     }
   }
 }
